Overview of all available checkFunctions

Produce an overview of all functions of class checkFunction available in the workspace or imported from packages. This overview includes the descriptions and a list of what classes the functions are each intended to be called on.

allCheckFunctions()

Value

An object of class functionSummary. This object has entries $name (the function names), $description (the function descriptions, as obtained from their description attributes) and $classes (the classes each function is indeded to be called on, as obtained from their classes attributes).

See also

Examples

allCheckFunctions()

#> 
#> ----------------------------------------------------------
#> name                      description                     
#> ------------------------- --------------------------------
#> identifyCaseIssues        Identify case issues            
#> 
#> identifyLoners            Identify levels with < 6 obs.   
#> 
#> identifyMissing           Identify miscoded missing       
#>                           values                          
#> 
#> identifyNums              Identify misclassified numeric  
#>                           or integer variables            
#> 
#> identifyOutliers          Identify outliers               
#> 
#> identifyOutliersTBStyle   Identify outliers (Turkish      
#>                           Boxplot style)                  
#> 
#> identifyWhitespace        Identify prefixed and suffixed  
#>                           whitespace                      
#> 
#> isCPR                     Identify Danish CPR numbers     
#> 
#> isEmpty                   Check if the variable contains  
#>                           only a single value             
#> 
#> isKey                     Check if the variable is a key  
#> 
#> isSingular                Check if the variable contains  
#>                           only a single value             
#> 
#> isSupported               Check if the variable class is  
#>                           supported by dataMaid.          
#> ----------------------------------------------------------
#> 
#> Table: Table continues below
#> 
#>  
#> -----------------------------
#> classes                      
#> -----------------------------
#> character, factor            
#> 
#> character, factor            
#> 
#> character, Date, factor,     
#> integer, labelled, logical,  
#> numeric                      
#> 
#> character, factor, labelled  
#> 
#> Date, integer, numeric       
#> 
#> Date, integer, numeric       
#> 
#> character, factor, labelled  
#> 
#> character, Date, factor,     
#> integer, labelled, logical,  
#> numeric                      
#> 
#> character, Date, factor,     
#> integer, labelled, logical,  
#> numeric                      
#> 
#> character, Date, factor,     
#> integer, labelled, logical,  
#> numeric                      
#> 
#> character, Date, factor,     
#> integer, labelled, logical,  
#> numeric                      
#> 
#> character, Date, factor,     
#> integer, labelled, logical,  
#> numeric                      
#> -----------------------------
#>