Semi-artificial data about masterpieces of art

A dataset with information about 200 painting and their painters. Each observation in the dataset corresponds to a painting. A single artificial variable, namely an artist ID variable, has been included. Otherwise the information should be truthful.

artData

Format

A data frame with 200 rows and 11 variables.

ArtistID

A unique ID used for cataloging the artists (fictional).

ArtistName

The name of the artist.

NoOfMiddlenames

The number of middlenames the artist has.

Title

The title of the painting.

Year

The approximate year in which the painting was made.

Location

The current location of the painting.

Continent

The continent of the current location of the painting.

Width

The width of the painting, in centimeters.

Height

The height of the painting, in centimers.

Media

The media/materials of the painting.

Movement

The artistic movement(s) the painting belongs to.

Source

Semi-artificial dataset constructed based on the Master Works of Art dataset available from Data Explorer.

Examples

data(artData)