Default checks for character variables

Default options for which checks to perform on character type variables in check and makeDataReport, possibly user-modified by adding extra function names using add or removing default function names with remove.

defaultCharacterChecks(remove = NULL, add = NULL)

Arguments

remove
Character vector of function names. Checks to remove from the returned vector

add
Character vector of function names. Checks to add to the returned vector

Value

A vector of function names.