A checkFunction for identifying case issues

A checkFunction to be called from check that identifies values in a vector that appear multiple times with different case settings.

identifyCaseIssues(v, nMax = 10)

Arguments

v
A character, factor, or labelled variable to check.

nMax
The maximum number of problematic values to report. Default is 10. Set to Inf if all problematic values are to be included in the outputted message, or to 0 for no output.

Value

A checkResult with three entires: $problem (a logical indicating whether case issues where found), $message (a message describing which values in v resulted in case issues) and $problemValues (the problematic values in their original format). Note that Only unique problematic values are listed and they are presented in alphabetical order.

See also

check, allCheckFunctions, checkFunction, checkResult

Examples

 identifyCaseIssues(c("val", "b", "1", "1", "vAl", "VAL", "oh", "OH"))

#> Note that there might be case problems with the following levels: OH, VAL, oh, vAl, val.