Set check arguments for makeDataReport

This function is a tool for easily specifying the checks argument of makeDataReport. Note that all available check function options can be inspected by calling allCheckFunctions().

setChecks(character = defaultCharacterChecks(),
  factor = defaultFactorChecks(), labelled = defaultLabelledChecks(),
  numeric = defaultNumericChecks(), integer = defaultIntegerChecks(),
  logical = defaultLogicalChecks(), Date = defaultDateChecks(),
  all = NULL)

Arguments

character
A character vector of function names to be used as checks for character variables. The default options are available by calling defaultCharacterChecks().

factor
A character vector of function names to be used as checks for factor variables. The default options are available by calling defaultFactorChecks().

labelled
A character vector of function names to be used as checks for labelled variables. The default options are available by calling defaultLabelledChecks().

numeric
A character vector of function names to be used as checks for numeric variables. The default options are available by calling defaultNumericChecks().

integer
A character vector of function names to be used as checks for integer variables. The default options are available by calling defaultIntegerChecks().

logical
A character vector of function names to be used as checks for logical variables. The default options are available by calling defaultLogicalChecks().

Date
A character vector of function names to be used as checks for Date variables. The default options are available by calling defaultDateChecks().

all
A character vector of function names to be used as checks for all variables. Note that this overrules the choices made for specific variable types by using the other arguments.

Value

A list with one entry for each data class supported by makeDataReport. Each entry then contains a character vector of function names that are to be called as checks for that variable type.

See also

makeDataReport, allCheckFunctions, defaultCharacterChecks, defaultFactorChecks, defaultLabelledChecks, defaultNumericChecks, defaultIntegerChecks, defaultLogicalChecks, defaultDateChecks

Examples

#Only identify missing values for characters, logicals and labelled variables:
  setChecks(character = "identifyMissing", factor = "identifyMissing",
     labelled = "identifyMissing")

#> $character
#> [1] "identifyMissing"
#> 
#> $factor
#> [1] "identifyMissing"
#> 
#> $labelled
#> [1] "identifyMissing"
#> 
#> $numeric
#> [1] "identifyMissing"  "identifyOutliers"
#> 
#> $integer
#> [1] "identifyMissing"  "identifyOutliers"
#> 
#> $logical
#> NULL
#> 
#> $Date
#> [1] "identifyOutliers" "identifyMissing" 
#> 

     
#Used in a call to makeDataReport():

# NOT RUN {
data(toyData)
makeDataReport(toyData, checks = setChecks(character = "identifyMissing",
   factor = "identifyMissing", labelled = "identifyMissing"), replace = TRUE)
# }