Extended example data to test the features of dataMaid

A dataset of constructed data used as test bed when using dataMaid for identifying potential errors in a dataset.

testData

Format

A data frame with 15 rows and 14 variables.

charVar

A character vector with a single missing observation.

factorVar

A factor vector with a miscoded missing observation, 999.

numVar

A numeric vector

intVar

An integer vector

boolVar

A logical vector with three missing observations.

keyVar

A character vector with unique codes for each observation.

emptyVar

A numeric vector where all entries are identical.

numOutlierVar

A numeric vector with a possible outlier (100).

smartNumVar

A numeric vector that takes only two different values.

cprVar

A character vector with levels in the format of Danish CPR numbers (social security numbers).

cprKeyVar

A character vector with levels in the format of Danish CPR numbers (social security numbers) with unique levels for each observation.

miscodedMissingVar

A character vector with levels corresponding to various miscoded (non-NA) misssing codes.

misclassifiedNumVar

A misclassified factor variable, where every level is a number and a many (12) different levels are in use.

dateVar

A Date vector.

labelledVar

A labelled vector with two missing observations.

Source

Artificial data

Examples

data(testData)