Summary function for original class

A summaryFunction type function, intended to be called from summarize, which finds the original class of a variable. This is just the class for all objects but those of class smartNum.

variableType(v, ...)

Arguments

v
A variable (vector).

...
Not in use.

Value

An object of class summaryResult with the following entries: $feature ("Variable type"), $result (the (original) class of v) and $value (identical to $result).

See also

summarize

Examples

 #For standard variables:
   varX <- c(rep(c(1,2,3), each=10))
   class(varX)

#> [1] "numeric"

   variableType(varX)

#> Variable type: numeric

 #For smartNum variables:
   smartX <- dataMaid:::smartNum(varX)
   class(smartX)

#> [1] "smartNum" "factor"  

   variableType(smartX)

#> Variable type: numeric