Semi-artificial data about the US presidents

A dataset with information about the first 45 US presidents as well as a 46th person, who is not a US president. The dataset was constructed to show the capabilities of dataMaid and therefore, it has been constructed to include errors and miscodings. Each observation in the dataset corresponds to a person. The dataset uses the non-standard class Name which is simply an attribute that has been added to two variables in order to show how dataMaid handles non-supported classes.

presidentData

Format

A data frame with 46 rows and 11 variables.

lastName

A Name type variable containing the last name of the president.

firstName

A Name type variable containing the first name of the president.

orderOfPresidency

A factor variable indicating the order of the presidents (with George Washington as number 1 and Donald Trump as number 45).

birthday

A Date variable with the birthday of the president

stateOfBirth

A character variable with the state in which the president was born.

assassinationAttempt

A numeric variable indicating whether there was an assassination attempt (1) or not (0) on the president.

sex

A factor variable with the sex of the president.

ethnicity

A factor variable with the ethnicity of the president.

presidencyYears

A numeric variable with the duration of the presidency, in years.

ageAtInauguration

A character variable with the age at inauguration.

favoriteNumber

A complex type variable with a fictional favorite number for each president.

Source

Artificial dataset constructed based on the US president dataset available from Data Explorer.

Examples

data(presidentData)